added logging for RoomService and RoomTableService

+++ src/main/java/com/example/demo/services/TimingService.java
@@ -140,7 +140,9 @@ public class TimingService {
      * @param timingConstraints The list of Timing objects to be deleted.
      */
     public void deleteTimingConstraints(List<Timing> timingConstraints){
-        timingRepository.deleteAll(timingConstraints);
+        if(timingConstraints != null){
+            timingRepository.deleteAll(timingConstraints);
+        }
     }
 
     /**
